delete ResponseListener interface
