+++ app/src/main/java/com/github/gogetters/letsgo/game/util/ogs/ResponseListener.kt
@@ -1,13 +1,21 @@
 package com.github.gogetters.letsgo.game.util.ogs
 
-interface ResponseListener<T> {
-    /**
-     * Sets the function to call on response
-     */
-    fun setOnResponse(action: (T) -> Unit)
+/**
+ * Listens to respond to http request
+ */
+class ResponseListener<T> {
+    private var action: (T) -> Unit = {}
+    private var response: T? = null
 
-    /**
-     * Function that is called when a response is received
-     */
-    fun onResponse(response: T)
+    fun setOnResponse(action: (T) -> Unit) {
+        this.action = action
+        if (response != null) {
+            action(response!!)
+        }
+    }
+
+    fun onResponse(response: T) {
+        this.response = response
+        action(response)
+    }
 }
